ai-engine/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
dashscope>=1.20.0
openai>=1.50.0
pydantic>=2.9.0
redis>=5.1.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
httpx>=0.27.0
pyyaml>=6.0
